Return ZK eventChannel part of the ZK store
In the open source library `github.com/docker/libkv`, the current implementation of the ZooKeeper store does not handle the connection to ZooKeeper dropping.

Returning a channel steaming ZooKeeper events as part of the ZK store can help to detect when such a disconnection happens, so a reconnection attempt can be performed.